pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-memory"
version = "0.1.0"
description = "Local-first, cognitively-inspired memory system for AI agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "sentence-transformers>=3.0.0",
    "open-clip-torch>=2.24.0",
    "qdrant-client>=1.9.0",
    "kuzu>=0.4.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["agent_memory*"]
